refactor(html-player): split controls bridge collaborators
