+++ libs/ui/playback/tsconfig.lib.json
@@ -8,5 +8,9 @@
         "types": []
     },
     "include": ["src/**/*.ts"],
-    "exclude": ["src/**/*.spec.ts", "src/**/*.test.ts"]
+    "exclude": [
+        "src/**/*.spec.ts",
+        "src/**/*.test.ts",
+        "src/**/*.spec-fixtures.ts"
+    ]
 }
